Extract config loading and host filtering options to reusable modules
